backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
openai
pydantic
